samples_requirements.txt
arrow>=1.2.2
requests>=2.22.0
importlib-metadata>=4.8.3
psycopg2-binary>=2.9.3
